1. Fix stream_pour_point_threshold_vector 2. Fix raster overlay with int64 type 3. Revert field_clipped_vector

+++ imWEBs/raster_extension.py
@@ -237,7 +237,7 @@ class RasterExtension:
             for col in range(cols):
                 if spatial1_ras[row, col] != no_data1 and spatial2_ras[row, col] != no_data2:
                     id = spatial1_ras[row, col] + spatial2_ras[row, col] * raster1_max
-                    temp[row,col]  = int(id)
+                    temp[row,col]  = np.int64(id)
         
         for row in range(rows):
             for col in range(cols):
